utils/docker: add friendly build helper script
So I wanted to fix a problem I found in the docs and wanted to build
them in Docker. The parts are all there but I made a helper script to
make it easier for people to get started.

This is a little more friendly than building via the Dockerfile in the
root of the repository; this way you don't need to copy files out of the
Docker container to view them in a local viewer.

It also runs CMake and `make` as a non-privileged user, which is good
for security.

Tested on macOS but should work on Linux systems also.

Passes `shellcheck` with no warnings/errors.